requirements.txt
ultralytics
roboflow
pandas
pillow
gradio